Remove LS support when dropping mails
Note that TLS is still used for the connection to the external SMTP server.
Removal of TLS is particularly useful for Go programs that refuse to connect
without TLS if TLS is available, making copying the snakeoil certificate into
the Go program’s container.  This is awkward, and brakes once the certificate
changes.

Besides, the hard-coded DNS name “postfix” has been the source of trouble
because sometimes, you have to contact the container under a different name.

If this turns out to be a bad idea, I have to do it completely differently,
e.g. by allowing to mount a certificate into the container.